Add unit test and rename files to kebab-case
